requirements.txt
SpeechRecognition
pyttsx3
Pyaudio
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib